+++ nixarr/default.nix
@@ -1,6 +1,8 @@
-vpnconfinement: {...}: {
+sub-merge: vpnconfinement: { pkgs, ... }: {
   imports = [
     vpnconfinement.nixosModules.default
     ./nixarr.nix
   ];
+
+  config.environment.systemPackages = [ sub-merge.packages."${pkgs.system}".default ];
 }